+++ ogn/parser/aprs_comment/lt24_parser.py
@@ -1,7 +1,7 @@
 import re
 
 from ogn.parser.utils import fpm2ms
-from ogn.parser.pattern import PATTERN_LT24_BEACON
+from ogn.parser.pattern import PATTERN_LT24_POSITION_COMMENT
 
 from .base import BaseParser
 
@@ -12,7 +12,7 @@ class LT24Parser(BaseParser):
 
     @staticmethod
     def parse_position(aprs_comment):
-        ac_match = re.search(PATTERN_LT24_BEACON, aprs_comment)
+        ac_match = re.search(PATTERN_LT24_POSITION_COMMENT, aprs_comment)
         return {'id': ac_match.group('id'),
                 'climb_rate': int(ac_match.group('climb_rate')) * fpm2ms if ac_match.group('climb_rate') else None,
                 'source': ac_match.group('source') if ac_match.group('source') else None}