+++ ogn/parser/aprs_comment/skylines_parser.py
@@ -1,7 +1,7 @@
 import re
 
 from ogn.parser.utils import fpm2ms
-from ogn.parser.pattern import PATTERN_SKYLINES_BEACON
+from ogn.parser.pattern import PATTERN_SKYLINES_POSITION_COMMENT
 
 from .base import BaseParser
 
@@ -12,6 +12,6 @@ class SkylinesParser(BaseParser):
 
     @staticmethod
     def parse_position(aprs_comment):
-        ac_match = re.search(PATTERN_SKYLINES_BEACON, aprs_comment)
+        ac_match = re.search(PATTERN_SKYLINES_POSITION_COMMENT, aprs_comment)
         return {'id': ac_match.group('id'),
                 'climb_rate': int(ac_match.group('climb_rate')) * fpm2ms if ac_match.group('climb_rate') else None}