+++ ogn/parser/aprs_comment/spider_parser.py
@@ -1,6 +1,6 @@
 import re
 
-from ogn.parser.pattern import PATTERN_SPIDER_BEACON
+from ogn.parser.pattern import PATTERN_SPIDER_POSITION_COMMENT
 
 from .base import BaseParser
 
@@ -11,7 +11,7 @@ class SpiderParser(BaseParser):
 
     @staticmethod
     def parse_position(aprs_comment):
-        ac_match = re.search(PATTERN_SPIDER_BEACON, aprs_comment)
+        ac_match = re.search(PATTERN_SPIDER_POSITION_COMMENT, aprs_comment)
         return {'id': ac_match.group('id'),
                 'signal_strength': int(ac_match.group('signal_strength')) if ac_match.group('signal_strength') else None,
                 'spider_id': ac_match.group('spider_id') if ac_match.group('spider_id') else None,