+++ ogn/parser/aprs_comment/spot_parser.py
@@ -1,6 +1,6 @@
 import re
 
-from ogn.parser.pattern import PATTERN_SPOT_BEACON
+from ogn.parser.pattern import PATTERN_SPOT_POSITION_COMMENT
 
 from .base import BaseParser
 
@@ -11,7 +11,7 @@ class SpotParser(BaseParser):
 
     @staticmethod
     def parse_position(aprs_comment):
-        ac_match = re.search(PATTERN_SPOT_BEACON, aprs_comment)
+        ac_match = re.search(PATTERN_SPOT_POSITION_COMMENT, aprs_comment)
         return {'id': ac_match.group('id'),
                 'model': int(ac_match.group('model')) if ac_match.group('model') else None,
                 'status': ac_match.group('status') if ac_match.group('status') else None}