+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # CHANGELOG
+## Unreleased
+- parser: Added support for OGNLT24 (LT24), OGSKYL (Skylines), OGSPID (Spider) and OGSPOT (Spot)
 
 ## 0.8.2: - 2018-01-20
 - parser: Better validation of timestamp, lat/lon and altitude